fixed wrong parameter in contract creation, some chore changes
